1、实现文件拖拽上传功能；2、实现文件下载功能

+++ src/components/minio-client/index.vue
@@ -29,6 +29,7 @@
         <el-form-item>
           <el-button type="primary" @click="resetSearch">重置搜索</el-button>
           <el-button type="primary" @click="refreshFiles">刷新数据</el-button>
+          <el-button type="primary" @click="createMenu">新增目录</el-button>
         </el-form-item>
         <el-form-item style="float: right" v-show="currentPath !== ''">
           <el-button type="primary" @click="back">返回</el-button>
@@ -38,7 +39,7 @@
     <div class="current-path">
       当前路径：<span> {{ activeBucket }}</span> / <span>{{ currentPath }}</span>
     </div>
-    <div class="file-box">
+    <div :class="['file-box', { 'drop-file-ing': dropFileIng }]" :id="randomId">
       <el-empty v-if="bucketFiles.length === 0">
         <el-button type="primary">上传文件</el-button>
       </el-empty>
@@ -70,9 +71,14 @@
     </div>
 
     <el-dialog title="新建Bucket" :visible.sync="createBucketDialogVisible">
-      <el-form :model="createForm">
+      <el-form :model="createForm" label-width="auto">
         <el-form-item label="Bucket名称">
-          <el-input v-model="createForm.name"></el-input>
+          <el-input v-model="createForm.name" placeholder="请填写Bucket名称,不能包含中文！"></el-input>
+        </el-form-item>
+        <el-form-item label="权限">
+          <el-select v-model="createForm.policy" placeholder="请选择Bucket权限">
+            <el-option :label="n.label" :value="n.value" v-for="n in bucketPolicy" :key="n.label"></el-option>
+          </el-select>
         </el-form-item>
       </el-form>
       <div slot="footer">
@@ -84,7 +90,18 @@
 </template>
 <script>
 const Minio = require('minio')
+// const mime = require('mime')
 const imgPrefix = ['png', 'jpg', 'gif', 'jpeg', 'jfif']
+const bucketPolicy = [
+  {
+    label: '私有的',
+    value: 'private'
+  },
+  {
+    label: '公开的',
+    value: 'public'
+  }
+]
 export default {
   name: 'minioClient',
   props: {
@@ -98,6 +115,7 @@ export default {
       searchForm: {
         key: ''
       },
+      randomId: new Date().getTime() + Math.random().toString(36).substr(2),
       pageLoading: false,
       connectFailed: true,
       minioClient: null,
@@ -111,9 +129,12 @@ export default {
       rightClickMenu: null,
       createBucketDialogVisible: false,
       createForm: {
-        name: ''
+        name: '',
+        policy: 'public'
       },
-      files: []
+      bucketPolicy,
+      files: [],
+      dropFileIng: false
     }
   },
   methods: {
@@ -123,10 +144,10 @@ export default {
       console.log('endPoint, port, useSSL, accessKey, secretKey', endPoint, port, useSSL, accessKey, secretKey)
       this.connectFailed = this.isEmpty(endPoint) || this.isEmpty(accessKey) || this.isEmpty(secretKey)
       if (this.connectFailed) return
-      if(port) port = parseInt(port)
+      if (port) port = parseInt(port)
       this.minioClient = new Minio.Client({
         endPoint: endPoint,
-        port: port ,
+        port: port,
         useSSL: useSSL || false,
         accessKey: accessKey,
         secretKey: secretKey
@@ -138,6 +159,7 @@ export default {
       this.currentPath = ''
       this.activeBucket = name
       this.getFileList(name)
+      console.log('name', typeof this.getPolicyCallBack)
     },
     showMenu(path) {
       this.currentPath += path
@@ -170,7 +192,6 @@ export default {
           let split = obj.name.split('.')
           if (imgPrefix.indexOf(split[split.length - 1]) > -1) {
             let imgPath = this.option.endPoint + ':' + this.option.port + '/' + this.activeBucket + '/'
-            console.log(this.option.endPoint.startsWith('http'))
             if (!this.option.endPoint.startsWith('http')) {
               imgPath = (this.option.useSSL ? 'https://' : 'http://') + imgPath
             }
@@ -241,9 +262,36 @@ export default {
         }
         this.buckets.push({ creationDate: new Date(), name: this.createForm.name })
         this.activeBucket = this.createForm.name
+        this.changeBucketPolicy(this.createForm)
         this.createForm.name = ''
+        this.createForm.policy = 'public'
         this.createBucketDialogVisible = false
       })
+    },
+    // 设置bucket 访问策略
+    async changeBucketPolicy({ name, policy }) {
+      const template = {
+        Version: '2012-10-17',
+        Statement:
+          policy === 'private'
+            ? []
+            : [
+                {
+                  Effect: 'Allow',
+                  Principal: { AWS: ['*'] },
+                  Action: ['s3:GetBucketLocation', 's3:ListBucket', 's3:ListBucketMultipartUploads'],
+                  Resource: ['arn:aws:s3:::' + name]
+                },
+                {
+                  Effect: 'Allow',
+                  Principal: { AWS: ['*'] },
+                  Action: ['s3:AbortMultipartUpload', 's3:DeleteObject', 's3:GetObject', 's3:ListMultipartUploadParts', 's3:PutObject'],
+                  Resource: ['arn:aws:s3:::' + name + '/*']
+                }
+              ]
+      }
+      const err = await this.minioClient.setBucketPolicy(name, JSON.stringify(template))
+      if (err) console.log('err', err)
     },
     menuRightMenu(event) {
       this.$contextmenu({
@@ -283,9 +331,20 @@ export default {
             }
           },
           {
-            label: '下载d',
+            label: '下载',
             onClick: () => {
-              this.downFile()
+              const el = document.createElement('a')
+              el.style.display = 'none'
+              el.setAttribute('target', '_blank')
+              const url = `${this.option.useSSL ? 'https://' : 'http://'}${this.option.endPoint}:${this.option.port}/${this.activeBucket}/${
+                this.currentPath
+              }${this.rightClickFile.name}`
+              console.log('path=' + url)
+              el.setAttribute('download', this.rightClickFile.name)
+              el.href = url
+              document.body.appendChild(el)
+              el.click()
+              document.body.removeChild(el)
             }
           }
         ],
@@ -295,17 +354,22 @@ export default {
         minWidth: 230 // 主菜单最小宽度
       })
     },
+    createMenu() {},
     removeMenu(prefix) {
-      this.pageLoading = true
-      const stream = this.minioClient.listObjects(this.activeBucket, this.currentPath + prefix, true)
-      stream.on('data', (obj) => {
-        this.removeFile(obj.name)
-      })
-      stream.on('end', () => {
-        this.back()
-        this.pageLoading = false
-        this.$message.success('删除目录成功！')
-      })
+      this.$confirm('无法撤回，是否要删除该目录？', '提示')
+        .then(() => {
+          this.pageLoading = true
+          const stream = this.minioClient.listObjects(this.activeBucket, this.currentPath + prefix, true)
+          stream.on('data', (obj) => {
+            this.removeFile(obj.name)
+          })
+          stream.on('end', () => {
+            this.back()
+            this.pageLoading = false
+            this.$message.success('删除目录成功！')
+          })
+        })
+        .catch(() => {})
     },
     downMenu() {},
     removeFile(fileName) {
@@ -319,7 +383,54 @@ export default {
         })
       })
     },
-    downFile() {}
+    downFile() {},
+    initDragUpload() {
+      const dropbox = document.getElementById(this.randomId)
+      dropbox.addEventListener('drop', this.enentDrop, false)
+      dropbox.addEventListener('dragleave', (e) => {
+        e.stopPropagation()
+        e.preventDefault()
+        this.dropFileIng = false
+      })
+      dropbox.addEventListener('dragenter', (e) => {
+        e.stopPropagation()
+        e.preventDefault
+        this.dropFileIng = true
+      })
+      dropbox.addEventListener('dragover', (e) => {
+        e.stopPropagation()
+        e.preventDefault()
+        this.dropFileIng = true
+      })
+    },
+    enentDrop(e) {
+      this.dropFileIng = false
+      e.stopPropagation()
+      e.preventDefault() //必填字段
+      let fileData = e.dataTransfer.files
+      for (let i = 0; i !== fileData.length; i++) {
+        const item = fileData[i]
+        const fr = new FileReader()
+        fr.readAsArrayBuffer(item) //读取文件内容,读取完成,result 属性中保存的将是被读取文件的 ArrayBuffer 数据对象.
+        fr.onload = async () => {
+          //文件读取成功回调
+          const dataUrl = Buffer.from(fr.result) // ArrayBuffer 转成 Buffer对象
+          console.log('dataUrl: ', typeof dataUrl)
+          await this.minioClient.putObject(this.activeBucket, this.currentPath + item.name, dataUrl, item.size, {
+            'Content-Type': item.type
+          })
+          this.refreshFiles()
+        }
+      }
+      this.$notify({
+        title: '上传成功',
+        message: `${fileData.length}个文件上传成功！`,
+        type: 'success'
+      })
+    }
+  },
+  mounted() {
+    this.initDragUpload()
   },
   created() {
     this.init()
@@ -336,6 +447,7 @@ export default {
 }
 .bucket-item {
   width: 23%;
+  /* padding: 0 10px; */
   height: 50px;
   border: 1px solid #f44336;
   border-radius: 8px;
@@ -345,6 +457,7 @@ export default {
   margin: 10px 0.948%;
   display: inline-block;
   cursor: pointer;
+  user-select: none;
   transition: all 0.3s ease-in-out;
 }
 .bucket-item > .bucket-item-title {
@@ -372,6 +485,23 @@ export default {
   flex-wrap: wrap;
   max-height: calc(100% - 170px);
   overflow: auto;
+  transition: all .3s ease-in-out;
+}
+.drop-file-ing.file-box::after {
+  content: '松开鼠标即可上传';
+  width: 100%;
+  height: 100%;
+  background: #0000005e;
+  position: absolute;
+  left: 0;
+  top: 0;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  color: aquamarine;
+  font-size: 37px;
+  font-weight: 600;
+  border: 2px solid #6adbab;
 }
 .file-item {
   width: 10%;
